Map parser is done

+++ src/math.rs
@@ -2,12 +2,12 @@ use std::ops::Add;
 
 #[derive(Debug, Copy, Clone)]
 pub struct Pos2 {
-    pub x: i16,
-    pub y: i16,
+    pub x: i32,
+    pub y: i32,
 }
 
 impl Pos2 {
-    pub fn new(x: i16, y: i16) -> Self {
+    pub fn new(x: i32, y: i32) -> Self {
         Self { x, y }
     }
 }
